// webapp/resources/sap/esh/search/ui/sinaNexTS/sina/AttributeType.ts
/*!
 * SAPUI5
 * Copyright (c) 2025 SAP SE or an SAP affiliate company. All rights reserved.
 * 
 */
export enum AttributeType {
    Double = "Double",
    Integer = "Integer",
    String = "String",
    ImageUrl = "ImageUrl",
    ImageBlob = "ImageBlob",
    GeoJson = "GeoJson", // Variants?
    Date = "Date", // Deprecated? Could use timestamp + a new format type instead
    Time = "Time", // Deprecated?
    Timestamp = "Timestamp",
    Group = "Group",
    INAV2_SearchTerms = "$$SearchTerms$$",
    INAV2_SuggestionTerms = "$$SuggestionTerms$$",
}